requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Mail==0.9.1
Flask-Bcrypt==1.0.1
Flask-SocketIO==5.3.5
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
Flask-Limiter==3.5.0
email-validator==2.0.0
Pillow==10.0.0
pyotp==2.9.0
qrcode==7.4.2
stripe==6.0.0
python-dotenv==1.0.0
cryptography==41.0.3
gunicorn==21.2.0
eventlet==0.33.3
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
SQLAlchemy==2.0.20
WTForms==3.0.1
bcrypt==4.0.1
python-socketio==5.9.0
dnspython==2.4.2
pypng==0.20220715.0
markdown==3.4.4
bleach==6.0.0
celery==5.3.4
PyJWT==2.8.0
requests==2.31.0
beautifulsoup4==4.12.2
python-slugify==8.0.1
Faker==19.3.1
pytest==7.4.2
coverage==7.3.1
docker==6.1.3

# OAuth Dependencies
Flask-Dance[sqla]>=6.0.0
blinker>=1.6.2
python-dotenv>=1.0.0
requests_oauthlib>=1.3.1

# Production dependencies
waitress==2.1.2
python-dotenv==1.0.0
gunicorn==21.2.0
sentry-sdk==1.39.1
prometheus-client==0.19.0
newrelic==9.3.0
